Use vim-kt instead of KaoriYa Vim
